celfiles_getdata {hiphop}	R Documentation

celfiles_getdata

Description

Import .CEL files

Usage

celfiles_getdata(celfiles_dir = NULL, cfv = NULL,
  gene_info = hiphop_gene_info, tag4_probe_info = hiphop_tag4_probe_info)

Arguments

celfiles_dir
Path to the directory with .CEL files to process

cfv
vector with .CEL file names (give basename with or without .CEL extension)

gene_info
Default data.frame hiphop_gene_info included with the package

tag4_probe_info
data.frame hiphop_tag4_probe_info included with the package

Details

The function processes raw .CEL files, breaks down data by probesets, uptags and downtags and adds gene information from YeastMine

Value

A list with objects named: all, all_standard, files + columns with short .CEL file names (4 last characters)

all

All data 318 x 318 = 101124 rows + Intensities from all arrays + annotation

all_standard

Subset of all containing the standard set of ORF (with bad tags excluded)

files

A table with the original .CEL filenames and the short name used in the output tables

[Package hiphop version 0.0.2 Index]