gene_info2_updated_20160903 {hiphop}	R Documentation

Information for features with an orf_on_chip present and mapped to current information from SGD as of 2016-09-03

Description

No selection for strains to be excluded (see tag4_probe_info2 for that purpose )

Usage

gene_info2_updated_20160903

Format

a data.frame with 6154 rows and 6 columns

column names

self describing

gene_name

gene name from SGD but whenever there was no gene name, the associated sgd_orf was used

Details

This table helps just with mapping orf_on_chip with sgd_orf, so updated information can be obtained

Information in part from: http://downloads.yeastgenome.org/curation/chromosomal_feature/SGD_features.tab

However, whenever there was no gene_name, the corresponding sgd_orf was used in place

[Package hiphop version 0.0.2 Index]