get_db_table {hiphop}	R Documentation

get_db_table

Description

get_db_table

Usage

get_db_table(dbcon, tblname, download.all = F, download.limit = 10000)

Arguments

dbcon
The database connection, e.g:

 ytdb_name = "yeast_track"
 yt_host = "yeast-track.pharmacy.ubc.ca"
 ytdb = connect_annotator(dbname = ytdb_name, host = yt_host)

 ycsdb_name = "ycs"
 ycs_host = "cn-websrv4.pharmacy.ubc.ca"
 ycs_db = connect_annotator(dbname = ycsdb_name, host = ycs_host)
 dbcon = ycs_db
 

tblname
A table name for a given database can be obtained with function: get_tables_info_from_database for example "analysis_info" is a table of the "yeast_track" database

download.all
Do you want to download all the data for the given table?, then set download.all = T, however, the number of rows may be very large and this will cause trouble, so the default is set to F, and a download.limit is set to a default of 10000 rows. If the table contains less than this limit, it will download all. You can set the download.limit value (see below)

download.limit
Default is 10000, however you can change it to another value greater than zero

Value

If download.all is set to T or download.all is set to F and there is a numeric value for download.limit greater than zero a table with all the data is obtained.

Important: If download.all is set to F and download.limit is set to NULL You get a table connection object that can be manipulated with dplyr functions

Example: Lets get AVG_G data for a given plate

platename ="14_07_15_OHC_Media_Replicates_Plate1_20150714_ps1_A_48_T_9.txt"
eod = get_plate_od(plate_name_in = platename)
# Plate Name is:  14_07_15_OHC_Media_Replicates_Plate1_20150714_ps1_A_48_T_9.txt 
# Plate Number is:  30285 

ytdb_name = "yeast_track"
yt_host = "yeast-track.pharmacy.ubc.ca"
ytdb = connect_annotator(dbname = ytdb_name, host = yt_host)

# # get table names and their column names from database connection
 ytdb_tables_info = get_tables_info_from_database(ytdb, with_colnames = T)

# names(ytdb_tables_info) : shows table: "analysis_info"
 ytdb_tables_info[["analysis_info"]] shows there is a column "Avg_G"

z = get_db_table(dbcon = ytdb,tblname = "analysis_info", download.all = F, download.limit = NULL)

library(dplyr)
avg_g = z 
# 48 x 4 table

[Package hiphop version 0.0.2 Index]