get_tables_info_from_database {hiphop}	R Documentation

get_tables_info_from_database

Description

Given a database connection, extracts all the table names and optionally get the column names for each table when using parameter 'with_colnames = T' (default)

Usage

get_tables_info_from_database(dbcon, with_colnames = T)

Arguments

dbcon
The database connection, e.g:

 ytdb_name = "yeast_track"
 yt_host = "yeast-track.pharmacy.ubc.ca"
 ytdb = connect_annotator(dbname = ytdb_name, host = yt_host)

 ycsdb_name = "ycs"
 ycs_host = "cn-websrv4.pharmacy.ubc.ca"
 ycs_db = connect_annotator(dbname = ycsdb_name, host = ycs_host)
 dbcon = ycs_db
 

with_colnames
If set to TRUE, extracts column names for each table

Value

A vector with all table names for the given database, if parameter 'with_colnames = F'

A named list where each element is a table and contains a vector of column names To get just the table names use function names() on the output

[Package hiphop version 0.0.2 Index]