hiphop {hiphop}	R Documentation

hiphop

Description

The hiphop package version 0.0.1 provides data and functions that facilitate analysis of hiphop data.

Gene annotation is consistent with YeastMine:

http://yeastmine.yeastgenome.org/yeastmine data update Nov 30, 2015

The cdf file for tag4 arrays (ftp://ftp.ebi.ac.uk/pub/databases/microarray/data/array/AFFY/A-AFFY-118/A-AFFY-118.cdf.tar.gz) does not define probesets but takes each probe as a probeset itself.

hiphop provides cdf environments for different subsets of probesets

Different objects (lists and data frames) are provided, you can type hiphop_ and hit tab to see the list of objects.

The meaning of the objects can be understood by their names and inspecting them.

Here is the list of objects and functions, so far included:

hiphop_fun_celfiles_getdata : Function, to see more information type: ?hiphop_fun_celfiles_getdata

hiphop_tag4_probe_info : Array structure

hiphop_gene_info: Gene Information

Below are objects with cdf environments for different subsets of probesets:

hiphop_customcdf

hiphop_hetess1095_besttags

hiphop_hetess1095_downtag

hiphop_hetess1095_up_down_tag

hiphop_hetess1095_uptag

hiphop_hom4810_besttags

hiphop_hom4810_downtag

hiphop_hom4810_up_down_tag

hiphop_hom4810_uptag

hiphop_raw_env

[Package hiphop version 0.0.2 Index]