process_prenorm {hiphop}	R Documentation

process_prenorm

Description

Processes the output of get_prenorm() to filter by the table given by Erica

Usage

process_prenorm(pout)

Arguments

pout
The output from get_prenorm()

Details

Erica's description of the table: It has all of the deleted and merged ORFs (as identified by SGD) removed, and half of the unassigned tags have been made 'uptags' and half 'downtags' which Guri agrees could be useful.

Value

Filtered get_prenorm() output: 16186 rows instead of 16269

[Package hiphop version 0.0.2 Index]