run_go_enrichment {hiphop}	R Documentation

Function to run GO enrichment analysis with web output of MCL clustered nodes

Description

The GO and Yeast Database information used here corresponds to September 03, 2016

Usage

run_go_enrichment(fdrThresh = 0.1, curr_exp, score)

Arguments

fdrThresh:
A threshhold for the FDR value (default 0.1)

curr_exp:
A name you give as a prefix for your output xgmml file

score:
the input data.frame. This data.frame should contain three columns named index, score and gene

index: Is a binary 0/1 representation of what is included as your selected group of genes(1) and what is not (0), all data provided will conform the universe

score: Typically is a log2Ratio

gene: gene names updated

Value

A list with three elements:

enrichMat_Ordered: A data.frame with results of the enrichment analysis ordered by increasing FDR (false discovery rate) and decreasing foldEnrichment

url_link: the link for visualization ( a pop-up window will also show a new tab at that url_link)

mcl_output: A list with four elements: xgmml, xgmml_file relate to the XML encoding and mcl.in and mcl.out are the input network and output clustered network from MCL

   Example (run this): Note, the object scoreMat_example is automatically available 
   
           myrun = run_go_enrichment(fdrThresh = 0.1, curr_exp = 'hiphop', score = scoreMat_example)
           
           You also will get the url printed on the screen and a pop-up window will open 
           (if it fails, your browser may ask you to trust the url and will give you a hint for how to accomplish that)

[Package hiphop version 0.0.2 Index]