scoreMat_example {hiphop}	R Documentation

Example data for enrichment analysis

Description

Data to be used as example for function 'run_go_enrichment', see documentation.

Usage

scoreMat_example

Format

A data frame with 4808 rows and 3 variables:

index

binary 0/1 representation of what is included as your selected group of genes(1) and what is not (0), all data provided will conform the universe

score

typically is a log2Ratio

gene

gene name

[Package hiphop version 0.0.2 Index]