tag4_probe_info2 {hiphop}	R Documentation

Information for the structure of the tag4 array with updated annotation for the sgd_orfs

Description

Information for the structure of the tag4 array with updated annotation for the sgd_orfs

Usage

tag4_probe_info2

Format

A data frame with 101124 rows and 18 columns

columns

18 columns self-describing :-)

exclusion_index_revised

index created by Mauricio for strains to exclude based on annotation from SGD, may change when I get new info from Erica

[Package hiphop version 0.0.2 Index]